backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.2
pymongo==4.6.1
pydantic-settings==2.2.1
pypdf==4.2.0
python-docx==1.1.2
numpy==2.4.3
scikit-learn==1.8.0
python-dotenv==1.0.1